Render BS3 dropdowns in wormhole
Moved common dropdown menu code to base, reverted BS3 tagname to back to ul
