+++ addon/components/base/bs-dropdown/menu.js
@@ -1,5 +1,6 @@
 import Component from '@ember/component';
 import { computed } from '@ember/object';
+import { notEmpty } from '@ember/object/computed';
 import layout from 'ember-bootstrap/templates/components/bs-dropdown/menu';
 
 /**
@@ -16,7 +17,7 @@ export default Component.extend({
   layout,
 
   /**
-   * Defaults to a `<ul>` tag in BS3 and a '<div>' tag in BS4. Change for other types of dropdown menus.
+   * Defaults to a `<ul>` tag in BS3 and a `<div>`` tag in BS4. Change for other types of dropdown menus.
    *
    * @property tagName
    * @type string
@@ -58,7 +59,7 @@ export default Component.extend({
   inNav: false,
 
   /**
-   * Applies only to BS4: by default the menu is rendered in the same place the dropdown. If you experience clipping
+   * By default the menu is rendered in the same place the dropdown. If you experience clipping
    * issues, you can set this to false to render the menu in a wormhole at the top of the DOM.
    *
    * @property renderInPlace
@@ -72,5 +73,59 @@ export default Component.extend({
     if (this.get('align') !== 'left') {
       return `dropdown-menu-${this.get('align')}`;
     }
-  })
+  }),
+  
+  isOpen: computed({
+    get() {
+      return false;
+    },
+    set(key, value) {
+      let update = this.get('_popperApi.update');
+      update && update();
+      return value;
+    }
+  }),
+
+  flip: true,
+
+  _popperApi: null,
+
+  inDom: notEmpty('toggleElement').readOnly(),
+
+  popperPlacement: computed('direction', 'align', function() {
+    let placement = 'bottom-start';
+    let { direction, align } = this.getProperties('direction', 'align');
+
+    if (direction === 'up') {
+      placement = 'top-start';
+      if (align === 'right') {
+        placement = 'top-end';
+      }
+    } else if (direction === 'left') {
+      placement = 'left-start';
+    } else if (direction === 'right') {
+      placement = 'right-start';
+    } else if (align === 'right') {
+      placement = 'bottom-end';
+    }
+    return placement;
+  }),
+
+  popperModifiers: computed('inNav', 'flip', function() {
+    return {
+      // @todo add offset config
+      applyStyle: {
+        enabled: !this.get('inNav')
+      },
+      flip: {
+        enabled: this.get('flip')
+      }
+    };
+  }),
+
+  actions: {
+    registerPopperApi(api) {
+      this.set('_popperApi', api);
+    }
+  }
 });
