+++ addon/components/bs3/bs-dropdown/menu.js
@@ -1,8 +1,5 @@
 import DropdownMenu from 'ember-bootstrap/components/base/bs-dropdown/menu';
 
 export default DropdownMenu.extend({
-  tagName: 'ul',
-
-  classNames: ['dropdown-menu'],
-  classNameBindings: ['alignClass']
+  tagName: 'ul'
 });
