--- addon/templates/components/bs3/bs-dropdown/menu.hbs
@@ -1,7 +1,0 @@
-{{yield
-  (hash
-    item=(component "bs-dropdown/menu/item")
-    link-to=(component "bs-dropdown/menu/link-to")
-    divider=(component "bs-dropdown/menu/divider")
-  )
-}}
